+++ config/processor.go
@@ -23,6 +23,13 @@ type Processor interface {
 	privateConfigProcessor()
 }
 
+// UnmarshalProcessor helper function to unmarshal a Processor config.
+// It checks if the config implements Unmarshallable and uses that if available,
+// otherwise uses Map.UnmarshalExact, erroring if a field is nonexistent.
+func UnmarshalProcessor(cfgMap *Map, cfg Processor) error {
+	return unmarshal(cfgMap, cfg)
+}
+
 // ProcessorSettings defines common settings for a component.Processor configuration.
 // Specific processors can embed this struct and extend it with more fields if needed.
 //
