+++ internal/configunmarshaler/testdata/duplicate-exporter.yaml
