+++ internal/configunmarshaler/testdata/duplicate-extension.yaml
