+++ internal/configunmarshaler/testdata/duplicate-pipeline.yaml
