+++ internal/configunmarshaler/testdata/duplicate-processor.yaml
