+++ internal/configunmarshaler/testdata/duplicate-receiver.yaml
