+++ internal/configunmarshaler/testdata/invalid-exporter-name-after-slash.yaml
