+++ internal/configunmarshaler/testdata/invalid-exporter-section.yaml
