+++ internal/configunmarshaler/testdata/invalid-exporter-type.yaml
