+++ internal/configunmarshaler/testdata/invalid-extension-name-after-slash.yaml
