+++ internal/configunmarshaler/testdata/invalid-extension-section.yaml
