+++ internal/configunmarshaler/testdata/invalid-extension-type.yaml
