+++ internal/configunmarshaler/testdata/invalid-pipeline-section.yaml
