+++ internal/configunmarshaler/testdata/invalid-pipeline-sub-config.yaml
