+++ internal/configunmarshaler/testdata/invalid-pipeline-type.yaml
