+++ internal/configunmarshaler/testdata/invalid-processor-section.yaml
