+++ internal/configunmarshaler/testdata/invalid-receiver-name-after-slash.yaml
