+++ internal/configunmarshaler/testdata/invalid-receiver-section.yaml
