+++ internal/configunmarshaler/testdata/invalid-receiver-sub-config.yaml
