+++ internal/configunmarshaler/testdata/invalid-receiver-type.yaml
