+++ internal/configunmarshaler/testdata/invalid-service-extensions-section.yaml
