+++ internal/configunmarshaler/testdata/unknown-exporter-type.yaml
