+++ internal/configunmarshaler/testdata/unknown-extension-type.yaml
