+++ internal/configunmarshaler/testdata/unknown-pipeline-type.yaml
