+++ internal/configunmarshaler/testdata/unknown-receiver-type.yaml
